% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_mapping.R
\docType{data}
\name{gene_mapping}
\alias{gene_mapping}
\title{Mapping gene names to Ensembl IDs}
\format{
A data frame with 59074 rows and 2 variables:
\describe{
  \item{ensembl}{Ensembl gene IDs}
  \item{name}{corresponding gene names}
}
}
\source{
\url{https://bioconductor.org/packages/release/data/annotation/html/EnsDb.Hsapiens.v79.html}
}
\usage{
gene_mapping
}
\description{
A data set showing the mapping between gene names and Ensembl gene IDs,
derived from the \pkg{EnsDb.Hsapiens.v79} Bioconductor package.
}
\keyword{datasets}
