% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{prox_fantope}
\alias{prox_fantope}
\title{Proximal operator on Fantope}
\usage{
prox_fantope(
  A,
  B,
  alpha,
  d,
  eps = 1e-05,
  inc = 1L,
  maxiter = 10L,
  verbose = 0L
)
}
\arguments{
\item{A}{A symmetric matrix.}

\item{B}{A symmetric matrix of the same dimension as \code{A}.}

\item{alpha}{Proximal parameter.}

\item{d}{Fantope parameter.}

\item{eps}{Precision of the result.}

\item{inc}{How many incremental eigenvalues to compute in each iteration.}

\item{maxiter}{Maximum number of iterations.}

\item{verbose}{Level of verbosity.}
}
\description{
This function solves the optimization problem
\deqn{\min\quad-tr(AX) + \frac{1}{2\alpha}||X - B||_F^2}{min  -tr(AX) + (0.5 / \alpha) * ||X - B||_F^2}
\deqn{s.t.\quad X\in \mathcal{F}^d}{s.t. X in F^d}
}
