% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_markers.R
\name{sort_markers}
\alias{sort_markers}
\title{Post-processing Selected Marker Genes}
\usage{
sort_markers(corr, markers)
}
\arguments{
\item{corr}{The sample correlation matrix, whose row and column names are gene names.}

\item{markers}{A list of marker genes. Each component of the list is a vector of marker
gene names corresponding to a cell type. All the gene names in this list
must appear in the row/column names of \code{corr}.}
}
\description{
This function reorders the selected marker genes using information of the sample
correlation matrix.
}
