% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refine_markers.R
\name{refine_markers}
\alias{refine_markers}
\title{Marker Gene Selection via Penalized Principal Component Analysis}
\usage{
refine_markers(
  mat_exp,
  range,
  markers,
  lambda,
  w = 1.5,
  thresh = 0.001,
  alpha = 0.01,
  maxit = 1000,
  eps = 1e-04,
  verbose = 0
)
}
\arguments{
\item{mat_exp}{The gene expression matrix in the original scale
(not logarithm-transformed), with rows standing for observations and
columns for genes. The matrix should include gene names as column names.}

\item{range}{A character vector of gene names, representing the range of genes in which
markers are sought.}

\item{markers}{A character vector of gene names giving the prior marker gene list.}

\item{lambda}{A tuning parameter to control the number of selected marker genes. A larger
value typically means a smaller number of genes.}

\item{w}{Tuning parameter to control the weight on prior information.
Larger \eqn{w} means genes not in the prior list are less likely
to be selected as markers.}

\item{thresh}{Below this threshold small factor loadings are treated as zeros.}

\item{alpha}{Step size of the optimization algorithm.}

\item{maxit}{Maximum number of iterations.}

\item{eps}{Tolerance parameter for convergence.}

\item{verbose}{Level of verbosity.}
}
\description{
This function refines a prior marker gene list by combining information from bulk
tissue data, based on the penalized principal component analysis. The current
implementation computes on one cell type at a time. To get marker genes for
multiple cell types, call this function iteratively.
}
